# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIMRS RSUD Dr. H. Chasan Boesoirie Ternate - API Documentation",
...
        "contact": {
            "name": "IT SIMRS RSUD Dr. H. Chasan Boesoirie",
            "url": "https://chasanboesoirie.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/service/medifirst2000",
            "description": "Server SIMRS (Proxy Anti-CORS - Aktif Otomatis)"
        },
        {
            "url": "https://chasanboesoirie.id/service/medifirst2000",
            "description": "Server Langsung (chasanboesoirie.id - Butuh Postman/No-CORS)"
        }
    ],
    "paths": {
...
        "/emr/get-cppt": {
            "get": {
                "tags": [
                    "CPPT Pasien (Catatan Perkembangan Terintegrasi)"
                ],
                "summary": "Daftar Catatan CPPT SOAP Pasien",
                "description": "Mengambil daftar catatan perkembangan pasien terintegrasi (CPPT) lengkap dengan rincian SOAP (Subjektif, Objektif, Asesmen, Planning), dokter/PPA penulis, ruangan perawatan, dan status verifikasi DPJP. Catatan: Gunakan parameter 'nocm' sebagai filter utama.",
                "operationId": "a64370f2df7b20c4385fce84aadcc79d",
                "parameters": [
                    {
                        "name": "nocm",
                        "in": "query",
                        "description": "Nomor Rekam Medis Pasien (No CM)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "0611944"
                        }
                    },
                    {
                        "name": "noregistrasifk",
                        "in": "query",
                        "description": "No Registrasi Kunjungan Pasien (Opsional)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "2609008289"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Catatan CPPT pasien berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "norec": {
                                                        "type": "string",
                                                        "example": "cppt-001239"
                                                    },
                                                    "nocppt": {
                                                        "type": "string",
                                                        "example": "1"
                                                    },
                                                    "tglinput": {
                                                        "type": "string",
                                                        "example": "2026-10-01 07:02:05"
                                                    },
                                                    "noregistrasi": {
                                                        "type": "string",
                                                        "example": "2609008289"
                                                    },
                                                    "nocm": {
                                                        "type": "string",
                                                        "example": "0611944"
                                                    },
                                                    "namapasien": {
                                                        "type": "string",
                                                        "example": "FATHAN ANDARA ADITYA AN"
                                                    },
                                                    "namaruangan": {
                                                        "type": "string",
                                                        "example": "ANAK"
                                                    },
                                                    "namalengkap": {
                                                        "description": "Nama Dokter / PPA pembuat CPPT",
                                                        "type": "string",
                                                        "example": "Dr. FAJRIAH A SUMADAYO"
                                                    },
                                                    "s": {
                                                        "description": "Subjektif (Keluhan Pasien)",
                                                        "type": "string",
                                                        "example": "Keluhan pasien / anamnesis"
                                                    },
                                                    "o": {
                                                        "description": "Objektif (Pemeriksaan Fisik & TTV)",
                                                        "type": "string",
                                                        "example": "Pemeriksaan fisik dan tanda vital"
                                                    },
                                                    "a": {
                                                        "description": "Asesmen (Analisis / Diagnosa Kerja)",
                                                        "type": "string",
                                                        "example": "Diagnosa kerja / asesmen"
                                                    },
                                                    "p": {
                                                        "description": "Planning (Rencana Tindakan & Terapi)",
                                                        "type": "string",
                                                        "example": "Rencana terapi dan instruksi"
                                                    },
                                                    "isverifikasi": {
                                                        "description": "Status Verifikasi DPJP",
                                                        "type": "boolean",
                                                        "example": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Inhuman"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Token otorisasi SIMRS Ternate. Default token: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJoaXMuamtuIn0.FhnuCCWQx9SOMwBcLx3NFMPH45-KUwkblNKQY9Debm26PJ7ygt4-1z7oSMrefb0qfJwKtp02kS1O0lupFcMz1Q",
                "name": "X-AUTH-TOKEN",
                "in": "header"
            }
        }
    },
...
}